requirements.txt
anyio==4.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asdfghjkl==0.1a2
asttokens==2.0.8
async-lru==2.0.4
attrs==23.1.0
autoray==0.5.1
Babel==2.12.1
backcall==0.2.0
backpack-for-pytorch==1.5.2
beautifulsoup4==4.12.2
bleach==6.0.0
certifi==2022.6.15.1
cffi==1.15.1
charset-normalizer==2.0.12
colorama==0.4.5
cramjam==2.7.0
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
einops==0.6.0
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.1.3
executing==1.1.1
fastjsonschema==2.18.0
fastparquet==2023.8.0
filelock==3.13.1
fonttools==4.37.1
fqdn==1.5.1
fsspec==2023.9.2
gpytorch==1.9.0
huggingface-hub==0.19.4
idna==3.3
imageio==2.22.1
importlib-metadata==6.8.0
ipykernel==6.16.1
ipython==8.5.0
isoduration==20.11.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==7.4.9
jupyter_core==5.3.1
jupyter_server==2.7.3
jupyter_server_terminals==0.4.4
jupyterlab==4.0.6
jupyterlab-pygments==0.2.2
jupyterlab_server==2.25.0
kiwisolver==1.4.4
laplace-torch==0.1a2
linear-operator==0.1.1
loguru==0.6.0
MarkupSafe==2.1.3
matplotlib==3.5.0
matplotlib-inline==0.1.6
mistune==3.0.1
nbclient==0.8.0
nbconvert==7.8.0
nbformat==5.9.2
nest-asyncio==1.5.6
notebook==7.0.4
notebook_shim==0.2.3
numpy==1.26.0
openpyxl==3.0.9
overrides==7.4.0
packaging==21.3
pandas==2.1.1
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==3.10.0
prometheus-client==0.17.1
prompt-toolkit==3.0.31
properscoring==0.1
psutil==5.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==13.0.0
pycparser==2.21
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2022.2.1
PyYAML==6.0.1
pyzmq==24.0.1
referencing==0.30.2
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.10.3
safetensors==0.4.0
scikit-learn==1.3.1
scipy==1.11.3
Send2Trash==1.8.2
setuptools-scm==7.0.5
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
stack-data==0.5.1
terminado==0.17.1
threadpoolctl==3.1.0
timm==0.9.11
tinycss2==1.2.1
tomli==2.0.1
torch==1.11.0
torchaudio==0.11.0
torchvision==0.12.0
tornado==6.2
tqdm==4.62.3
traitlets==5.10.0
typing_extensions==4.3.0
tzdata==2023.3
uri-template==1.3.0
urllib3==1.26.12
wcwidth==0.2.5
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.3
wget==3.2
win32-setctime==1.1.0
xlrd==2.0.1
zipp==3.17.0

